requirements.txt
Flask>=2.0
opencv-python-headless==4.9.0.80
opencv-contrib-python-headless==4.9.0.80
mysql-connector-python>=8.0
numpy<2.0
Pillow>=9.0
werkzeug>=2.0
